[BULK EDIT OF MEASURES] Refactoring of persistance part of code - WIP
[TRELLO CARD](https://trello.com/c/dMAf97m1/434-edit-measures-functionality-not-working-eg-edit-the-measure-with-end-date-or-regulation-6)
