+++ app/views/measures/bulks/_table_top.html.erb
@@ -1,6 +1,6 @@
 <div class="table-top">
   <div class="row">
-    <% if workbasket_author? && params[:submitted].blank? %>
+    <% if workbasket_is_editable? %>
       <div class="col-md-3">
         <button-dropdown :disabled="noSelectedMeasures">
           <template slot="text">&ndash; select bulk action &ndash;</template>
