+++ app/views/measures/bulks/edit.html.slim
@@ -10,14 +10,10 @@
       = link_to "Back to search results", measures_url(search_code: workbasket.initial_search_results_code)
 
 h2.heading-large
-  - if params[:submitted].present?
+  - if workbasket.awaiting_cross_check?
     | Submitted for Cross-Check
   - else
     = workbasket_author? ? 'Edit measures' : 'Review measures'
-
-/p
-  a
-    | Return to search results / amend selection
 
 script
   == "window.__workbasket_id = #{workbasket.id};"
